ensure that box left and right edge are float64
--HG--
branch : yt

+++ yt/data_objects/static_output.py
@@ -824,9 +824,9 @@ class Dataset(object):
         # list or other non-array iterable before calculating
         # the center
         if not isinstance(left_edge, np.ndarray):
-            left_edge = np.array(left_edge)
+            left_edge = np.array(left_edge, dtype='float64')
         if not isinstance(right_edge, np.ndarray):
-            right_edge = np.array(right_edge)
+            right_edge = np.array(right_edge, dtype='float64')
         c = (left_edge + right_edge)/2.0
         return self.region(c, left_edge, right_edge, **kwargs)
